Split the vulkan backend into multiple files.
